Use universal types for ints and longs

+++ lib/fst/interface/fsthash.h
@@ -54,7 +54,7 @@ public:
 	 * \param blockHash if true, use a multi-threaded blocked hash format (hash of hashes)
 	 * \return hash value of the input buffer
 	 */
-	unsigned long long HashBlob(unsigned char* blobSource, unsigned long long blobLength, bool blockHash = true) const
+	uint64_t HashBlob(unsigned char* blobSource, uint64_t blobLength, bool blockHash = true) const
 	{
 		return HashBlobSeed(blobSource, blobLength, FST_HASH_SEED, blockHash);
 	}
@@ -67,7 +67,7 @@ public:
   * \param blockHash if true, use a multi-threaded blocked hash format (hash of hashes)
 	* \return hash value of the input buffer
    */
-	unsigned long long HashBlobSeed(unsigned char* blobSource, unsigned long long blobLength, unsigned long long seed, bool blockHash = true) const
+	uint64_t HashBlobSeed(unsigned char* blobSource, uint64_t blobLength, uint64_t seed, bool blockHash = true) const
 	{
     if (!blockHash)
     {
@@ -83,11 +83,11 @@ public:
 		}
 
 		// block size to use for hashing has a lower bound for compression efficiency
-		unsigned long long minBlock = std::max(static_cast<unsigned long long>(HASH_SIZE),
+		uint64_t minBlock = std::max(static_cast<uint64_t>(HASH_SIZE),
 			1 + (blobLength - 1) / PREV_NR_OF_BLOCKS);
 
 		// And a higher bound for hasher compatability
-		unsigned int blockSize = static_cast<unsigned int>(std::min(minBlock, static_cast<unsigned long long>(INT_MAX)));
+		unsigned int blockSize = static_cast<unsigned int>(std::min(minBlock, static_cast<uint64_t>(INT_MAX)));
 
 		int nrOfBlocks = static_cast<int>(1 + (blobLength - 1) / blockSize);
 		nrOfThreads = std::min(nrOfThreads, nrOfBlocks);
@@ -95,7 +95,7 @@ public:
 		unsigned int lastBlockSize = 1 + (blobLength - 1) % blockSize;
 		float blocksPerThread = static_cast<float>(nrOfBlocks) / nrOfThreads;
 
-		unsigned long long* blockHashes = new unsigned long long[nrOfBlocks];
+		uint64_t* blockHashes = new uint64_t[nrOfBlocks];
 
 #pragma omp parallel num_threads(nrOfThreads)
 		{
@@ -132,7 +132,7 @@ public:
 
 		}  // end parallel region and join all threads
 
-    unsigned long long allBlockHash = blockHashes[0];
+    uint64_t allBlockHash = blockHashes[0];
 
     // combine multiple hashes
     if (nrOfBlocks > 1)
